pyproject.toml
[project]
name = "ramanujan-formulas"
version = "1.0.0"
description = "Autonomous Mathematical Discovery via Genetic Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.55",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "mpmath>=1.3.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "numpy>=2.3.5",
    "sympy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
